+++ src/state_class.js
@@ -13,16 +13,18 @@ export default class State {
     this.level = {
       fillColor: -1,
       strokeColor: -1,
-      fontColor: -1
+      fontColor: -1,
     };
     this.currentColors = {};
     this.originalColors = {};
     this.originalValue = this.xgraph.getValueCell(mxcell);
     this.currentValue = this.originalValue;
-    this.originalLink = mxcell.getAttribute('link');
-    this.currentLink = mxcell.getAttribute('link');
+    let link = this.xgraph.getLink(mxcell);
+    if (link === undefined) link = null;
+    this.originalLink = link;
+    this.currentLink = link;
 
-    this.styles.forEach(style => {
+    this.styles.forEach((style) => {
       const color = this.xgraph.getStyleCell(mxcell, style);
       this.currentColors[style] = color;
       this.originalColors[style] = color;
@@ -30,9 +32,9 @@ export default class State {
   }
 
   setState(rule, serie) {
-    u.log(1, 'state_class.setState()');
-    u.log(0, 'state_class.setState() Rule', rule);
-    u.log(0, 'state_class.setState() Serie', serie);
+    u.log(1, 'State.setState()');
+    u.log(0, 'State.setState() Rule', rule);
+    u.log(0, 'State.setState() Serie', serie);
     if (rule.matchSerie(serie)) {
       const shapeMaps = rule.getShapeMaps();
       const textMaps = rule.getTextMaps();
@@ -40,9 +42,10 @@ export default class State {
       const value = rule.getValueForSerie(serie);
       const FormattedValue = rule.getFormattedValue(value);
       const level = rule.getThresholdLevel(value);
+
       // SHAPE
       let cellProp = this.getCellProp(rule.data.shapeProp);
-      shapeMaps.forEach(shape => {
+      shapeMaps.forEach((shape) => {
         if (!shape.isHidden() && shape.match(cellProp)) {
           this.matchedShape = true;
           this.matched = true;
@@ -51,29 +54,42 @@ export default class State {
             if (rule.toColorize(value)) {
               this.setColorStyle(rule.data.style, rule.getColorForValue(value));
             }
+            this.overlayIcon = rule.toIconize(value);
           }
         }
       });
+
       // TEXT
       cellProp = this.getCellProp(rule.data.textProp);
-      textMaps.forEach(text => {
+      textMaps.forEach((text) => {
         if (!text.isHidden() && text.match(cellProp)) {
           this.matchedText = true;
           this.matched = true;
           if (this.globalLevel <= level) {
-            this.setText(rule.getReplaceText(this.originalValue, FormattedValue));
+            if (rule.toValorize(value)) {
+              this.setText(rule.getReplaceText(this.originalValue, FormattedValue));
+            } else {
+              // Hide text
+              this.setText(rule.getReplaceText(this.originalValue, ''));
+            }
           }
         }
       });
       // LINK
       cellProp = this.getCellProp(rule.data.linkProp);
-      linkMaps.forEach(link => {
+      linkMaps.forEach((link) => {
         if (!link.isHidden() && link.match(cellProp)) {
           this.matchedLink = true;
           this.matched = true;
+          if (this.globalLevel <= level) {
+            if (rule.toLinkable(value)) {
+              this.currentLink = rule.getLink();
+            }
+          }
         }
       });
     }
+    u.log(0, 'State.setState() state', this);
   }
 
   unsetState() {
@@ -86,6 +102,7 @@ export default class State {
     this.matchedShape = false;
     this.matchedText = false;
     this.matchedLink = false;
+    u.log(0, 'State.unsetState() state', this);
   }
 
   getCellProp(prop) {
@@ -104,7 +121,7 @@ export default class State {
   }
 
   unsetColor() {
-    this.styles.forEach(style => {
+    this.styles.forEach((style) => {
       this.unsetColorStyle(style);
     });
   }
@@ -122,7 +139,7 @@ export default class State {
   }
 
   unsetLevel() {
-    this.styles.forEach(style => {
+    this.styles.forEach((style) => {
       this.unsetLevelStyle(style);
     });
     this.globalLevel = -1;
@@ -200,24 +217,35 @@ export default class State {
     u.log(1, 'State.applyState()');
     if (this.matched) {
       if (this.matchedShape) {
-        this.styles.forEach(style => {
+        this.styles.forEach((style) => {
+          // Apply colors
           this.xgraph.setStyleCell(this.mxcell, style, this.getCurrentColorStyle(style));
         });
+        // Apply icons
+        if (this.overlayIcon) {
+          this.xgraph.addOverlay(this.getTextLevel(), this.mxcell);
+        } else {
+          this.xgraph.removeOverlay(this.mxcell);
+        }
       }
       if (this.matchedText) {
         this.xgraph.setValueCell(this.mxcell, this.getCurrentText());
       }
-      // TODO:LINK
+      if (this.matchedLink) {
+        this.xgraph.addLink(this.mxcell, this.currentLink);
+      }
     } else this.restoreCell();
   }
 
   restoreCell() {
     this.unsetState();
-    this.styles.forEach(style => {
+    this.styles.forEach((style) => {
       this.xgraph.setStyleCell(this.mxcell, style, this.getCurrentColorStyle(style));
     });
     this.xgraph.setValueCell(this.mxcell, this.getCurrentText());
-    this.mxcell.setAttribute('link', this.getCurrentLink());
+    // this.mxcell.setAttribute('link', this.getCurrentLink());
+    this.xgraph.removeOverlay(this.mxcell);
+    this.xgraph.addLink(this.mxcell, this.originalLink);
   }
 
   prepare() {